% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_matrix.R
\name{corr_matrix}
\alias{corr_matrix}
\title{Correlation Matrix 
download "Season_Stats_NBA.csv" and assign as nba_data before using function.
This function allows you to compute the correlation matrix.}
\usage{
corr_matrix(y)
}
\description{
Correlation Matrix 
download "Season_Stats_NBA.csv" and assign as nba_data before using function.
This function allows you to compute the correlation matrix.
}
\examples{
corr_matrix()
}
\keyword{correlation}
