% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_pts.R
\name{max_pts_year}
\alias{max_pts_year}
\title{Player with the highest points
download "Season_Stats_NBA.csv" and assign as nba_data before using function.
This function allows you to find the player with the highest points for a given year}
\usage{
max_pts_year(y)
}
\description{
Player with the highest points
download "Season_Stats_NBA.csv" and assign as nba_data before using function.
This function allows you to find the player with the highest points for a given year
}
\examples{
max_pts()
}
\keyword{correlation}
