% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oldest_player.R
\name{oldest_player}
\alias{oldest_player}
\title{Oldest player in the league
download "Season_Stats_NBA.csv" and assign as nba_data before using function.
This function allows you to find the player the oldest player in the league for that year}
\usage{
oldest_player(y)
}
\description{
Oldest player in the league
download "Season_Stats_NBA.csv" and assign as nba_data before using function.
This function allows you to find the player the oldest player in the league for that year
}
\examples{
oldest_player()
}
\keyword{oldest}
